+++ i8253.cs
@@ -62,7 +62,7 @@ internal class i8253 : Device
         mappings[0x0043] = this;
     }
 
-    public override (ushort, bool) IO_Read(ushort port)
+    public override (byte, bool) IO_Read(ushort port)
     {
         if (port == 0x0040)
             return (GetCounter(0), _timers[0].is_pending);
@@ -76,16 +76,16 @@ internal class i8253 : Device
         return (0xaa, false);
     }
 
-    public override bool IO_Write(ushort port, ushort value)
+    public override bool IO_Write(ushort port, byte value)
     {
         if (port == 0x0040)
-            LatchCounter(0, (byte)value);
+            LatchCounter(0, value);
         else if (port == 0x0041)
-            LatchCounter(1, (byte)value);
+            LatchCounter(1, value);
         else if (port == 0x0042)
-            LatchCounter(2, (byte)value);
+            LatchCounter(2, value);
         else if (port == 0x0043)
-            Command((byte)value);
+            Command(value);
 
         return _timers[0].is_pending || _timers[1].is_pending || _timers[2].is_pending;
     }
